priority to pip venv + refactor setup
