+++ .flake8
@@ -1,0 +1,9 @@
+[flake8]
+ignore =
+    E501
+    E203
+    W503
+max-line-length = 79
+max-complexity = 15
+per-file-ignores =
+    __init__.py:F401
