--- .github/workflows/black.yml
@@ -1,10 +1,0 @@
-name: Lint black
-
-on: [push, pull_request]
-
-jobs:
-  lint:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - uses: psf/black@stable
