+++ .github/workflows/python-app.yml
@@ -21,9 +21,9 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install flake8 pytest
+        pip install flake8 black
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Lint with flake8
       run: flake8 . --count --show-source --statistics
-    - name: Test with pytest
-      run: unittest discover .
+    - name: Test with unittest
+      run: python -m unittest discover -s ./test
