+++ .gitignore
@@ -2,3 +2,4 @@
 data/*
 __pycache__
 *.egg-info*
+.venv
