--- environment.yml
@@ -1,78 +1,0 @@
-name: dymfile
-channels:
-  - conda-forge
-dependencies:
-  - blosc=1.21.4=hc338f07_0
-  - bzip2=1.0.8=h3422bc3_4
-  - c-ares=1.19.1=hb547adb_0
-  - ca-certificates=2023.5.7=hf0a4a13_0
-  - cffi=1.15.1=py311hae827db_3
-  - cfgv=3.3.1=pyhd8ed1ab_0
-  - cftime=1.6.2=py311h4add359_1
-  - curl=8.1.2=h912dcd9_0
-  - distlib=0.3.6=pyhd8ed1ab_0
-  - filelock=3.12.0=pyhd8ed1ab_0
-  - hdf4=4.2.15=h8111dcc_6
-  - hdf5=1.14.0=nompi_h6b85c65_103
-  - icu=72.1=he12128b_0
-  - identify=2.5.24=pyhd8ed1ab_0
-  - krb5=1.20.1=h69eda48_0
-  - libaec=1.0.6=hb7217d7_1
-  - libblas=3.9.0=16_osxarm64_openblas
-  - libcblas=3.9.0=16_osxarm64_openblas
-  - libcurl=8.1.2=h912dcd9_0
-  - libcxx=16.0.4=h4653b0c_0
-  - libedit=3.1.20191231=hc8eb9b7_2
-  - libev=4.33=h642e427_1
-  - libexpat=2.5.0=hb7217d7_1
-  - libffi=3.4.2=h3422bc3_5
-  - libgfortran=5.0.0=12_2_0_hd922786_31
-  - libgfortran5=12.2.0=h0eea778_31
-  - libiconv=1.17=he4db4b2_0
-  - libjpeg-turbo=2.1.5.1=h1a8c8d9_0
-  - liblapack=3.9.0=16_osxarm64_openblas
-  - libnetcdf=4.9.2=nompi_h0a2dbf5_105
-  - libnghttp2=1.52.0=hae82a92_0
-  - libopenblas=0.3.21=openmp_hc731615_3
-  - libsqlite=3.42.0=hb31c410_0
-  - libssh2=1.10.0=h7a5bd25_3
-  - libxml2=2.11.4=he3bdae6_0
-  - libzip=1.9.2=h76ab92c_1
-  - libzlib=1.2.13=h03a7124_4
-  - llvm-openmp=16.0.4=h1c12783_0
-  - lz4-c=1.9.4=hb7217d7_0
-  - ncurses=6.3=h07bb92c_1
-  - netcdf4=1.6.3=nompi_py311h6706b3f_102
-  - nodeenv=1.8.0=pyhd8ed1ab_0
-  - numpy=1.24.3=py311hb8f3215_0
-  - openssl=3.1.1=h53f4e23_1
-  - packaging=23.1=pyhd8ed1ab_0
-  - pandas=2.0.2=py311h9e438b8_0
-  - pip=23.1.2=pyhd8ed1ab_0
-  - platformdirs=3.5.1=pyhd8ed1ab_0
-  - pre-commit=3.3.2=pyha770c72_0
-  - pycparser=2.21=pyhd8ed1ab_0
-  - python=3.11.3=h1456518_0_cpython
-  - python-dateutil=2.8.2=pyhd8ed1ab_0
-  - python-tzdata=2023.3=pyhd8ed1ab_0
-  - python_abi=3.11=3_cp311
-  - pytz=2023.3=pyhd8ed1ab_0
-  - pyyaml=6.0=py311he2be06e_5
-  - readline=8.2=h92ec313_1
-  - setuptools=67.7.2=pyhd8ed1ab_0
-  - six=1.16.0=pyh6c4a22f_0
-  - snappy=1.1.10=h17c5cce_0
-  - tk=8.6.12=he1e0b03_0
-  - typing-extensions=4.6.2=hd8ed1ab_0
-  - typing_extensions=4.6.2=pyha770c72_0
-  - tzdata=2023c=h71feb2d_0
-  - ukkonen=1.0.1=py311hd6ee22a_3
-  - virtualenv=20.23.0=pyhd8ed1ab_0
-  - wheel=0.40.0=pyhd8ed1ab_0
-  - xarray=2023.5.0=pyhd8ed1ab_0
-  - xz=5.2.6=h57fd34a_0
-  - yaml=0.2.5=h3422bc3_2
-  - zstd=1.5.2=hf913c23_6
-  - pip:
-      - dymfile==0.1
-prefix: /Users/ash/miniforge3/envs/dymfile
