--- setup.cfg
@@ -1,28 +1,0 @@
-[metadata]
-name = dymfile
-# If you want to publish on PyPI then the following are also required
-url = 
-author = Jules Lehodey
-author_email = 
-# This is useful but not required
-version = 0.1
-
-# If you are using the "src" layout then we need the following
-# https://setuptools.pypa.io/en/latest/userguide/declarative_config.html
-[options]
-package_dir=
-    =src
-packages=find:
-
-[options.packages.find]
-where=src
-
-[flake8]
-ignore =
-    E501
-    E203
-    W503
-max-line-length = 79
-max-complexity = 15
-per-file-ignores =
-    __init__.py:F401
